Allow multiple tilda instances with d-bus
This change adds support for multiple tilda
instance which each their own d-bus interface.

This change removes the generated code and directly adds
a modified version of the generated D-Bus code to the
source tree.

The reason is that to support multiple d-bus interfaces
we need to use different bus names, object names and interface names.
However, gdbus-codegen requires a hard coded interface name in
its interface definition file.

The code added to the source tree is based on a modified version
of the code generated by gdbus-codegen which has modified the interface
functions to allow dynamically generated object and interface names.